Replace `methods` array by an object.
As documented in issue #224, `methods` field value could be replaced by
an object intead of an array.

Quoting @BelfordZ: "it's naturally preserve the uniqueness" of names
through object keys. Furthermore this uniqueness cannot be enforced with
the meta-schema if it's an array.

In this commit I replaced the array by an object and I forwarded the
name contraint with `patternProperties` keyword.